requirements.txt
discord.py
yt-dlp
spotdl
